pyproject.toml
[project]
name = "ycla-ai"
version = "0.1.0"
description = "Chatting api"
requires-python = ">=3.11"
dependencies = [
    "aioredis>=2.0.1",
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "azure-core>=1.34.0",
    "azure-search-documents>=11.5.2",
    "celery[redis]>=5.5.2",
    "docx2txt>=0.9",
    "eventlet>=0.40.0",
    "fastapi>=0.115.12",
    "gunicorn>=23.0.0",
    "langchain>=0.3.23",
    "langchain-community>=0.3.23",
    "langchain-openai>=0.3.16",
    "loguru>=0.7.3",
    "openai>=1.73.0",
    "psycopg2>=2.9.10",
    "pycurl>=7.45.6",
    "pydantic>=2.11.3",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.0",
    "redis>=5.3.0",
    "shortuuid>=1.0.13",
    "sqlalchemy>=2.0.40",
    "sqlmodel>=0.0.24",
    "tiktoken>=0.9.0",
    "uvicorn>=0.34.0",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
